+++ smile/src/test/java/com/fasterxml/jackson/dataformat/smile/mapper/NumberBeanTest.java
@@ -75,7 +75,7 @@ public class NumberBeanTest extends BaseTestForSmile
             g.writeEndObject();
         }
         result = MAPPER.readValue(bytes.toByteArray(), NumberWrapper.class);
-        assertEquals(Integer.valueOf(123), result.nr);
+        assertEquals(123, result.nr);
 
         bytes = new ByteArrayOutputStream();
         try (SmileGenerator g = smileGenerator(bytes, true)) {
@@ -84,7 +84,7 @@ public class NumberBeanTest extends BaseTestForSmile
             g.writeEndObject();
         }
         result = MAPPER.readValue(bytes.toByteArray(), NumberWrapper.class);
-        assertEquals(Long.valueOf(Long.MAX_VALUE), result.nr);
+        assertEquals(Long.MAX_VALUE, result.nr);
 
         bytes = new ByteArrayOutputStream();
         try (SmileGenerator g = smileGenerator(bytes, true)) {
@@ -108,7 +108,7 @@ public class NumberBeanTest extends BaseTestForSmile
             g.writeEndObject();
         }
         result = MAPPER.readValue(bytes.toByteArray(), NumberWrapper.class);
-        assertEquals(Float.valueOf(0.25f), result.nr);
+        assertEquals(0.25f, result.nr);
 
         bytes = new ByteArrayOutputStream();
         try (SmileGenerator g = smileGenerator(bytes, true)) {
@@ -117,7 +117,7 @@ public class NumberBeanTest extends BaseTestForSmile
             g.writeEndObject();
         }
         result = MAPPER.readValue(bytes.toByteArray(), NumberWrapper.class);
-        assertEquals(Double.valueOf(0.5), result.nr);
+        assertEquals(0.5, result.nr);
 
         bytes = new ByteArrayOutputStream();
         try (SmileGenerator g = smileGenerator(bytes, true)) {
@@ -151,13 +151,13 @@ public class NumberBeanTest extends BaseTestForSmile
 
             assertToken(JsonToken.VALUE_NUMBER_INT, p.nextToken());
             assertEquals(NumberType.INT, p.getNumberType());
-            assertEquals(Integer.valueOf(101), p.getNumberValue());
-            assertEquals(Integer.valueOf(101), p.getNumberValueExact());
+            assertEquals(101, p.getNumberValue());
+            assertEquals(101, p.getNumberValueExact());
 
             assertToken(JsonToken.VALUE_NUMBER_FLOAT, p.nextToken());
             assertEquals(NumberType.DOUBLE, p.getNumberType());
-            assertEquals(Double.valueOf(0.25), p.getNumberValue());
-            assertEquals(Double.valueOf(0.25), p.getNumberValueExact());
+            assertEquals(0.25, p.getNumberValue());
+            assertEquals(0.25, p.getNumberValueExact());
 
             assertToken(JsonToken.VALUE_NUMBER_INT, p.nextToken());
             assertEquals(NumberType.LONG, p.getNumberType());
@@ -166,8 +166,8 @@ public class NumberBeanTest extends BaseTestForSmile
 
             assertToken(JsonToken.VALUE_NUMBER_FLOAT, p.nextToken());
             assertEquals(NumberType.FLOAT, p.getNumberType());
-            assertEquals(Float.valueOf(0.5f), p.getNumberValue());
-            assertEquals(Float.valueOf(0.5f), p.getNumberValueExact());
+            assertEquals(0.5f, p.getNumberValue());
+            assertEquals(0.5f, p.getNumberValueExact());
 
             assertToken(JsonToken.VALUE_NUMBER_INT, p.nextToken());
             assertEquals(NumberType.BIG_INTEGER, p.getNumberType());
